+++ src/ApolloClient.ts
@@ -53,10 +53,6 @@ import {
 } from './data/extensions';
 
 import {
-  QueryTransformer,
-} from './queries/queryTransform';
-
-import {
   MutationBehavior,
   MutationBehaviorReducerMap,
   MutationQueryReducersMap,
@@ -100,7 +96,7 @@ export default class ApolloClient {
   public initialState: any;
   public queryManager: QueryManager;
   public reducerConfig: ApolloReducerConfig;
-  public queryTransformer: QueryTransformer;
+  public addTypename: boolean;
   public resultTransformer: ResultTransformer;
   public resultComparator: ResultComparator;
   public shouldForceFetch: boolean;
@@ -146,24 +142,26 @@ export default class ApolloClient {
     reduxRootSelector,
     initialState,
     dataIdFromObject,
-    queryTransformer,
     resultTransformer,
     resultComparator,
     ssrMode = false,
     ssrForceFetchDelay = 0,
     mutationBehaviorReducers = {} as MutationBehaviorReducerMap,
+    addTypename = true,
+    queryTransformer,
   }: {
     networkInterface?: NetworkInterface,
     reduxRootKey?: string,
     reduxRootSelector?: string | ApolloStateSelector,
     initialState?: any,
     dataIdFromObject?: IdGetter,
-    queryTransformer?: QueryTransformer,
     resultTransformer?: ResultTransformer,
     resultComparator?: ResultComparator,
     ssrMode?: boolean,
     ssrForceFetchDelay?: number
     mutationBehaviorReducers?: MutationBehaviorReducerMap,
+    addTypename?: boolean,
+    queryTransformer?: any,
   } = {}) {
     if (reduxRootKey && reduxRootSelector) {
       throw new Error('Both "reduxRootKey" and "reduxRootSelector" are configured, but only one of two is allowed.');
@@ -175,6 +173,11 @@ export default class ApolloClient {
           'please use the "reduxRootSelector" instead.'
       );
       this.reduxRootKey = reduxRootKey;
+    }
+
+    if (queryTransformer) {
+      throw new Error('queryTransformer option no longer supported in Apollo Client 0.5. ' +
+        'Instead, there is a new "addTypename" option, which is on by default.');
     }
 
     if (!reduxRootSelector && reduxRootKey) {
@@ -193,7 +196,7 @@ export default class ApolloClient {
     this.initialState = initialState ? initialState : {};
     this.networkInterface = networkInterface ? networkInterface :
       createNetworkInterface({ uri: '/graphql' });
-    this.queryTransformer = queryTransformer;
+    this.addTypename = addTypename;
     this.resultTransformer = resultTransformer;
     this.resultComparator = resultComparator;
     this.shouldForceFetch = !(ssrMode || ssrForceFetchDelay > 0);
@@ -439,7 +442,7 @@ export default class ApolloClient {
       networkInterface: this.networkInterface,
       reduxRootSelector: reduxRootSelector,
       store,
-      queryTransformer: this.queryTransformer,
+      addTypename: this.addTypename,
       resultTransformer: this.resultTransformer,
       resultComparator: this.resultComparator,
       reducerConfig: this.reducerConfig,