+++ test/graphqlSubscriptions.ts
@@ -163,6 +163,7 @@ describe('GraphQL Subscriptions', () => {
     // This test calls directly through Apollo Client
     const client = new ApolloClient({
       networkInterface: network,
+      addTypename: false,
     });
 
     const sub = client.subscribe(options).subscribe({
@@ -189,6 +190,7 @@ describe('GraphQL Subscriptions', () => {
       networkInterface: network,
       reduxRootSelector: (state: any) => state.apollo,
       store: createApolloStore(),
+      addTypename: false,
     });
 
     const obs = queryManager.startGraphQLSubscription(realOptions);
@@ -227,6 +229,7 @@ describe('GraphQL Subscriptions', () => {
       networkInterface: network,
       reduxRootSelector: (state: any) => state.apollo,
       store: createApolloStore(),
+      addTypename: false,
     });
 
     const sub = queryManager.startGraphQLSubscription(realOptions).subscribe({
@@ -271,6 +274,7 @@ describe('GraphQL Subscriptions', () => {
       networkInterface: network,
       reduxRootSelector: (state: any) => state.apollo,
       store: createApolloStore(),
+      addTypename: false,
     });
 
     const observableQuery = queryManager.watchQuery({